Full API

Index:

MagicTensors.MagicTensors

MagicTensors.AbstractCAMPS

MagicTensors.AbstractCliffordGate

MagicTensors.AbstractCliffordGateSet

MagicTensors.AbstractCliffordUnitary

MagicTensors.AbstractDisentangler

MagicTensors.AbstractDmrgDisentangler

MagicTensors.AbstractPauli

MagicTensors.AbstractPauliSum

MagicTensors.AbstractStandAloneDisentangler

MagicTensors.AbstractSweepDisentangler

MagicTensors.GreedySweepDisentangler

MagicTensors.IterativeDisentangler

MagicTensors.IterativeDmrgDisentangler

MagicTensors.QubitCAMPS

MagicTensors.QubitCliffordGate

MagicTensors.QubitCliffordGateSet

MagicTensors.QubitCliffordUnitary

MagicTensors.QubitPauli

MagicTensors.QubitPauliSum

MagicTensors.QubitPauliSum

MagicTensors.TrivialDisentangler

MagicTensors.apply!

MagicTensors.apply_and_disentangle!

MagicTensors.apply_to_clifford!

MagicTensors.apply_to_clifford_dagger!

MagicTensors.apply_to_mps!

MagicTensors.conjugate

MagicTensors.disentangle!

MagicTensors.dmrg!

MagicTensors.embed

MagicTensors.entanglement_entropy

MagicTensors.expectation

MagicTensors.get_clifford

MagicTensors.get_clifford_copy

MagicTensors.get_mps

MagicTensors.nsites

MagicTensors.stabilizer_entropy

MagicTensors.transform!

Docstrings:

MagicTensors.MagicTensors — Module

MagicTensors

TODO

MagicTensors.AbstractCAMPS — Type

AbstractCAMPS

Interface for Clifford Augmented Matrix Product States (CAMPS).

Description

AbstractCAMPS defines an supertype for Clifford Augmented Matrix Product States (CAMPS). A CAMPS is a ansatz for a pure quantum state of the form |ψ⟩ = C |MPS⟩, where C is a Clifford unitary, and |MPS⟩ is a matrix product state [TODO].

Required methods

For a CAMPS x implementations must support:

apply_to_mps!(x, args...; kwargs...): – applies args..., with kwargs... to the MPS part of the CAMPS.

apply_to_clifford!(x, args...; kwargs...: – applies args..., with kwargs... to the Clifford part C of the CAMPS.

apply_to_clifford_dagger!(x, args...; kwargs...: – applies args..., with kwargs... to the daggered Clifford part C† of the CAMPS.

get_clifford(x): – returns (by reference) the Clifford unitary C, which should be of the type <:AbstractCliffordUnitary.

get_clifford_copy(x): – returns a copy of the Clifford unitary C, which should be of the type <:AbstractCliffordUnitary.

get_mps(x): – returns (by reference) the matrix product state part of the CAMPS as an ITensorMPS.MPS object.

References

TODO

See also

QubitCAMPS, AbstractCliffordUnitary

apply!, apply_to_clifford!, apply_to_clifford_dagger!, apply_to_mps!, get_clifford, get_clifford_copy, get_mps

MagicTensors.AbstractCliffordGate — Type

AbstractCliffordGate

Abstract supertype for few-qubit/qudit Clifford gates that can act both within the stabilizer formalism and as explicit unitary matrices in the computational basis.

Description

AbstractCliffordGate defines an abstract interface for Clifford gates that act on a few qubit or qudit systems. While Clifford gates are in principle Clifford unitaries, they are not a subtype of AbstractCliffordUnitary because they fulfil a different role in the logic of the MagicTensors package. AbstractCliffordGate objects constitute the bridge between tensor networks and the stabilizer formalism because they can act on a Clifford unitary (stabilizer formalism) and they can return their explicit unitary matrix for use in tensor networks. In practice, the latter restricts the employability of AbstractCliffordGate objects to only a few sites.

Required methods

For a Clifford gates x, y and compatible Clifford unitary C, implementations must support:

Base.:(==)(x,y) -> x==y: – returns true if the gates x and y are identical, false otherwise.

Base.Matrix(x) -> M: – returns the gate x in the form of an explicit complex matrix M.

apply!(C, x, sites) -> C': – applies gate x to C at sites sites in-place, i.e. it replaces C → x C.

nsites(x) -> n: – returns the number n of qubits on which the gate x acts.

Notes

Types that implement this interface should be immutable.

See also

QubitCliffordGate

apply!, nsites

MagicTensors.AbstractCliffordGateSet — Type

AbstractCliffordGateSet

Abstract supertype for a list of AbstractCliffordGate objects.

Description

AbstractCliffordGateSet defines an abstract interface for an iterable and enumerated list of Clifford gates (each acting on the same number of sites). Clifford gate sets are useful, for example, for brute force searches for disentangling Clifford circuits.

Required methods

For a Clifford gate set x implementations must support:

Base.getindex(x, i) -> gᵢ: – returns the ith Clifford gate gᵢ in the set x.

Base.length(x) -> K: – returns the number of gates K in the set x.

Base.one(x) -> I: – return the identity gate w.r.t. set x; the identity I need not be in x.

nsites(x) -> n: – returns the number of sites n on which the gates in the set act.

See also

AbstractCliffordGate, QubitCliffordGateSet

nsites

MagicTensors.AbstractCliffordUnitary — Type

AbstractCliffordUnitary

Abstract supertype for Clifford unitaries.

Description

AbstractCliffordUnitary defines an abstract interface for Clifford unitaries that operate on qubits, qudits, or potentially other quantum systems. Clifford unitaries map Pauli strings (AbstractPauli) to Pauli strings under conjugation. Implementations of AbstractCliffordUnitary should function even for a large system sizes by employing the stabilizer formalism.

Required methods

For a Clifford unitarys C and D, and compatible Pauli P (each acting on n sites) implementations must support:

ConcreteCliffordUnitary(n) -> Iₙ: – constructor for the identity unitary as a Clifford unitary, where ConcreteCliffordUnitary<:AbstractCliffordUnitary.

Base.(==)(x,y) -> x==y: – returns true if the Pauli strings x and y are identical, false otherwise.

Base.:*(C, D) -> C D: – returns the product C D of Clifford unitaries C and D.

Base.inv(C) -> C⁻¹: – returns the inverse of C.

apply!(D, C): – applies C to D in-place, i.e. subsitutes D → C D.

conjugate(P, C) -> Q: – returns a Pauli Q = C P C† with the same type as P.

nsites(C) -> n: – returns the number of sites n.

See also

AbstractPauli, QubitCliffordUnitary

conjugate, nsites

MagicTensors.AbstractDisentangler — Type

AbstractDisentangler

TODO

MagicTensors.AbstractDmrgDisentangler — Type

AbstractDmrgDisentangler <: AbstractDisentangler

TODO

MagicTensors.AbstractPauli — Type

AbstractPauli

Abstract supertype for qubit/qudit Pauli strings that act on a system of qubits or qudits.

Description

AbstractPauli defines an abstract interface for Pauli strings acting on qubit or qudit systems.

Required methods

Implementations ConcretePauli<:AbstractPauli must support

ConcretePauli(n) -> x: – returns an identity Pauli string x on n sites.

Base.(==)(x,y) -> x==y: – returns true if the Pauli strings x and y are identical, false otherwise.

Base.≈(x,y) -> x≈y: – returns true if the Pauli strings x and y are the same, ignoring their phases, false otherwise.

Base.-(x) -> -x: – returns the Pauli string -x.

embed(x, m, sites) -> y: – returns an Pauli string on m≥n sites, where the Pauli on site i in x is put to site site[i] in y.

nsites(x) -> n: – returns the number of sites n the Pauli string acts on.

Additional methods

Base.:*(x, y) -> x*y: – returns the product of two Pauli strings.

See also

QubitPauli

embed, nsites

MagicTensors.AbstractPauliSum — Type

AbstractPauliSum

Abstract supertype for weighted unordered lists of Pauli strings.

Description

AbstractPauliSum defines an abstract interface for weighted unordered lists of Pauli strings. A Pauli sum x represents the sum over K Pauli strings Pᵢ weighted by αᵢ, i.e. x = ∑ αᵢ Pᵢ.

Required methods

Concrete implimentations ConcretePauliSum<:AbstractPauliSum must support:

ConcretPauliSum(n): – constructor returning and empty (K=0) Pauli sum on n sites.

Base.length(x) -> K: – returns the number of Pauli strings K in the sum.

Base.push!(x, P, α): – adds a Pauli string P with the given coefficient α to the sum.

Base.isapprox(x,y; kwargs...) -> x≈y: – returns true if the Pauli strings x and y are approximately equal (with kwargs... piped into the scalar isapprox function); false otherwise.

Base.(==)(x,y) -> x==y: – returns true if the Pauli sums x and y are equal, false otherwise.

embed(x, m, sites) -> y: – returns an Pauli sum on m≥n sites, where each Pauli on site i in x is put to site site[i] in y.

nsites(x) -> n: – returns the number n of qubits/qudits on which the Pauli strings in the sum act.

ITensorMPS.MPO(x, sites) -> mpo: – returns the Pauli sum as an ITensorMPS.MPO on the site indices sites.

Additional methods

For Pauli sums x, y, and scalar α it may be necessary for some applications to also implement the following methods:

Simple arithmethic: -x, x+y, x-y, α*x, and x/α, where x and y may also be plain QubitPauli objects.

Base.:*(x, y) -> x*y: – returns the product of two Pauli sums.

LinearAlgebra.adjoint(x) -> x†: – returns the adjoint of the Pauli sum.

LinearAlgebra.conj(x) -> x̅: – returns the complex conjugate of the Pauli sum.

LinearAlgebra.transpose(x) -> xᵀ: – returns the transpose of the Pauli sum.

LinearAlgebra.ishermitian(x): – returns true if the Pauli sum is Hermitian, and false otherwise.

See also

QubitPauliSum

embed, nsites

MagicTensors.AbstractStandAloneDisentangler — Type

AbstractStandAloneDisentangler <: AbstractDisentangler

TODO

MagicTensors.AbstractSweepDisentangler — Type

AbstractSweepDisentangler <: AbstractStandAloneDisentangler

TODO

MagicTensors.GreedySweepDisentangler — Type

GreedySweepDisentangler

TODO

MagicTensors.IterativeDisentangler — Type

IterativeDisentangler

TODO

MagicTensors.IterativeDmrgDisentangler — Type

IterativeDmrgDisentangler

TODO

MagicTensors.QubitCAMPS — Type

QubitCAMPS <: CAMPS

Implementation of AbstractCAMPS for qubit systems.

See also

AbstractCAMPS

MagicTensors.QubitCliffordGate — Type

QubitCliffordGate <: AbstractCliffordGate

Implementation of AbstractCliffordGate for qubits using the QuantumClifford package.

See also

AbstractCliffordGate

MagicTensors.QubitCliffordGateSet — Type

QubitCliffordGateSet{K, S} <: AbstractCliffordGateSet

Implementation of AbstractCliffordGateSet for K qubits using the QuantumClifford package. The symbol S specifies the gate set variant:

QubitCliffordGateSet{K, :empty}: The empty K-qubit Clifford gates set.

QubitCliffordGateSet{K, :all}: All K-qubit Clifford gates without phases.

QubitCliffordGateSet{K, :full}: All K-qubit Clifford gates with 4^K phase variants.

QubitCliffordGateSet{2, :entangle}: 19 two-qubit Clifford gates with entangling properties, excluding the identity.

See also

AbstractCliffordGateSet

MagicTensors.QubitCliffordUnitary — Type

QubitCliffordUnitary <: AbstractCliffordUnitary

Implementation of AbstractCliffordUnitary for qubit systems.

See also

AbstractCliffordUnitary

MagicTensors.QubitPauli — Type

QubitPauli <: AbstractPauli

Implementation of AbstractPauli for qubit systems using the QuantumClifford package.

QubitPauli objects are immutable with only a single field pauli::QuantumClifford.PauliOperator.

Constructors

QubitPauli(pauli::QuantumClifford.PauliOperator): – return the QubitPauli for a given QuantumClifford.PauliOperator.

QubitPauli(n::Int): – return an identity operator on n qubits.

@QubitPauli_str(s::AbstractString): – macro, using the same String to Pauli conversion as in QuantumClifford (e.g. QubitPauli"-iIXYZ").

See also

AbstractPauli

MagicTensors.QubitPauliSum — Type

QubitPauliSum <: AbstractPauliSum

Implementation of AbstractPauliSum for qubit systems using the QuantumClifford package.

See also

AbstractPauliSum

MagicTensors.QubitPauliSum — Method

QubitPauliSum(n::Int)

Constructor for QubitPauliSum that initializes an empty (complex) sum for a given number of qubits n.

MagicTensors.TrivialDisentangler — Type

TrivialDisentangler

TODO

MagicTensors.apply! — Function

apply!

TODO

MagicTensors.apply_and_disentangle! — Function

apply_and_disentangle!

TODO

MagicTensors.apply_to_clifford! — Function

apply_to_clifford!

TODO

MagicTensors.apply_to_clifford_dagger! — Function

apply_to_clifford_dagger!

TODO

MagicTensors.apply_to_mps! — Function

apply_to_mps!

TODO

MagicTensors.conjugate — Function

conjugate(P, C) -> Q

Returns the conjugation of an Pauli string or Pauli sum by a Clifford unitary.

Description

Computes Q = C P C†, where C is a Clifford unitary and P is a Pauli string or Pauli sum. The result Q is a Pauli string or Pauli sum of the same type as P.

Arguments

P::Union{AbstractPauli, AbstractPauliSum}: – A Pauli string or Pauli sum to be conjugated.

C::AbstractCliffordUnitary: – A Clifford unitary to conjugate with.

Returns

Q::Union{AbstractPauli, AbstractPauliSum}: – The conjugated Pauli string or Pauli sum C P C†.

See also

AbstractCliffordUnitary, AbstractPauli, AbstractPauliSum

MagicTensors.disentangle! — Function

disentangle!

TODO

MagicTensors.dmrg! — Function

dmrg!

TODO

MagicTensors.embed — Function

embed

TODO

MagicTensors.entanglement_entropy — Function

entanglement_entropy

TODO

MagicTensors.expectation — Function

expectation

TODO

MagicTensors.get_clifford — Function

get_clifford

TODO

MagicTensors.get_clifford_copy — Function

get_clifford_copy

TODO

MagicTensors.get_mps — Function

get_mps

TODO

MagicTensors.nsites — Function

nsites(obj) -> n

Returns the number of qubit or qudit sites n in an object.

Description

Returns the number of qubits or qudits in an object obj.

Objects obj may be any of the following types: - AbstractCAMPS - AbstractCliffordGate - AbstractCliffordGateSet - AbstractCliffordUnitary - AbstractPauli - AbstractPauliSum - ITensorMPS.MPS

Arguments

obj: – Object acting on or representing a certain number of qubits or qudits.

Returns

n::Int: – Number of qubits or qudits in object obj.

See also

AbstractCAMPS, AbstractCliffordGate, AbstractCliffordGateSet, AbstractCliffordUnitary, AbstractPauli, AbstractPauliSum,

MagicTensors.stabilizer_entropy — Function

stabilizer_entropy

TODO

MagicTensors.transform! — Function

transform!

TODO: write docstring.